Bucky Wagon Updates
Motor controller logic changed for UQM motor, was previously EATON motor
controller. With proper startup sequence. Also the cooling logic was
updated to allow more parameters that are being measured to effect fan
and pump speed. The logic was also updated so that an EMP fan controller
could be run.