Regen and Power steering Updated and CCP handler
Updated the regen limits to increase max total potential regen capbility
of the motor. Was previously limited to 240 nm. Also updated power
steering logic so that it has the potential for % speed control and also
comes on when brake pedal is pressed so that way at a stop the pump
continues to run. Also added CCP handler so that the code generates an
A2L file to interface with controller in other calibration software
other than MotoTune like ATI Vision.